pyproject.toml
[project]
name = "vivpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.2",
    "pre-commit>=4.0.1",
    "pydantic-settings>=2.6.0",
    "ruff>=0.6.9",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    "python-multipart>=0.0.12",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "trio>=0.27.0",
    "pytest-asyncio>=0.24.0",
]
